LookUpTable

Documentation

data LookUpTable

empty :: LookUpTable

Returns a LookUpTable that maps from any key -> Nothing

get :: LookUpTable -> String -> Maybe Int

Given a LookUpTable and a key, returns Just val if it finds a mapping or else Nothing

set :: LookUpTable -> String -> Int -> LookUpTable

Given a LookUpTable, a key and a val, creates a new LookUpTable with the additional mapping