Main

Documentation

main :: IO LookUpTable

Entry point. You can choose to execute with a file, or use as a line interpreter and it executes the instructions

executeFromFile :: IO LookUpTable

executeFromTerminal :: IO LookUpTable

readFileToString :: IO [Char]